Moved maintainer configuration to a separate file
